🚧 0SFMS7 task: split evaluator review helpers
